src/sdhypermodern_python/wikipedia.py
# src/sdhypermodern-python/wikipedia.py

# standard library
from dataclasses import dataclass

# third-party packages
import click
import desert
import marshmallow
import requests

# local packages


API_URL = "https://{language}.wikipedia.org/api/rest_v1/page/random/summary"


@dataclass
class Page:
    title: str
    extract: str


schema = desert.schema(Page, meta={"unknown": marshmallow.EXCLUDE})


def random_page(language: str = "en") -> Page:
    url = API_URL.format(language=language)

    try:
        with requests.get(url, timeout=15) as response:
            response.raise_for_status()
            data = response.json()
            return schema.load(data)
    except (requests.RequestException, marshmallow.ValidationError) as error:
        message = str(error)
        raise click.ClickException(message) from error
